requirements.txt
flet>=0.21.1
mysql-connector-python>=8.0.26
cryptography>=42.0.5  # for secure password handling 